Pin down versions of development dependencies
This change was prompted by the tests failing when using VCR 3.0.1 with
Webmock 2.0.2, because support for Webmock 2+ wasn't added until VCR
3.0.2. By being explicit about the versions of the development
dependencies we avoid wasting time trying to figure out why tests pass
on one machine but not another.

I've used the `~>` operator here so that we still get patch versions
that are higher than the one specified, but not lower.